refactor(server): extract env-readers into persistence/env-readers.ts (part of #1559) by claude
Moves readOptionalPositiveNumber, readPositiveInteger, readNonNegativeInteger, and
readBooleanFlag helpers out of persistence.ts into a dedicated module. These are pure
string-parsing utilities used by the MySQL config resolver.

Also wires persistence/index.ts to re-export mysql-tables, types, and env-readers so
the subdirectory is consumable as a single entry point.

No runtime behavior change. Refs #1559.
