+++ apps/server/src/persistence/index.ts
@@ -1,1 +1,3 @@
 export * from "./mysql-tables";
+export * from "./types";
+export * from "./env-readers";
